feat: small handle to prevent reloading in wsl vm

+++ flake.nix
@@ -75,13 +75,14 @@
                 ${pkgs.nixos-rebuild}/bin/nixos-rebuild test --flake . --use-remote-sudo --verbose --option eval-cache false --show-trace
             esac
 
-            if [ $? -eq 0 ]; then
+            if [ $? -eq 0 -a $HOSTNAME != "nixos" ]; then
               hyprctl reload
               pkill waybar
               unset GDK_BACKEND && waybar > /dev/null 2>&1 &
               disown
-              ${pkgs.coreutils}/bin/kill -INT $$ # simulates ^C
             fi
+
+            ${pkgs.coreutils}/bin/kill -INT $$ # simulates ^C
           '')
         ];
       };
